requirements.txt
Django==1.9.1
djangorestframework==3.3.2
pip==7.1.2
setuptools==19.2
wheel==0.26.0
pytz==2015.7
Pillow==3.1.0